admin can now add users

+++ src/app/components/user/user-edit/user-edit.component.ts
@@ -13,10 +13,11 @@ export class UserEditComponent implements OnInit {
   @ViewChild('f') form: NgForm;
 
   userId: string;
-  user: any;
+  user = {};
   errorFlag = false;
   errorMsg = 'You have an error';
   type = '';
+  newUser = false;
 
   constructor(private userService: UserService, private activatedRoute: ActivatedRoute, private router: Router) {
   }
@@ -26,27 +27,36 @@ export class UserEditComponent implements OnInit {
       .subscribe(
         (params: any) => {
           this.userId = params['uid'];
-          this.userService.findUserById(this.userId).subscribe((data: any) => {
-            this.type = data.type;
-            this.user = data;
-          });
+          if (this.userId !== 'new') {
+            this.userService.findUserById(this.userId).subscribe((data: any) => {
+              this.type = data.type;
+              this.user = data;
+            });
+          } else {
+            this.newUser = true;
+          }
         }
       );
   }
 
   submit() {
     this.userService.findUserByUsername(this.form.value.username).subscribe((data: any) => {
-      if (data != null && data._id !== this.userId) {
+      if ((this.newUser && data != null && data._id !== this.userId) || (!this.newUser && data != null)) {
         this.errorFlag = true;
         this.errorMsg = 'That username is already taken';
         return;
       }
-      this.user['username'] = this.form.value.username;
-      this.user['firstName'] = this.form.value.firstName;
-      this.user['lastName'] = this.form.value.lastName;
-      this.user['type'] = this.form.value.type;
-      this.userService.updateUser(this.userId, this.user).subscribe(() => {
-      });
+      if (this.newUser) {
+        this.userService.registerByAdmin(this.form.value.username, this.form.value.password, this.form.value.type).subscribe(() => {
+        });
+      } else {
+        this.user['username'] = this.form.value.username;
+        this.user['firstName'] = this.form.value.firstName;
+        this.user['lastName'] = this.form.value.lastName;
+        this.user['type'] = this.form.value.type;
+        this.userService.updateUser(this.userId, this.user).subscribe(() => {
+        });
+      }
       this.errorMsg = 'User Successfully Updated!';
       this.errorFlag = true;
       return this.router.navigate(['/users']);
@@ -54,9 +64,10 @@ export class UserEditComponent implements OnInit {
   }
 
   onDelete() {
-    this.userService.deleteUser(this.userId).subscribe(() => {
-    });
+    if (!this.newUser) {
+      this.userService.deleteUser(this.userId).subscribe(() => {
+      });
+    }
     return this.router.navigate(['/users']);
   }
-
 }
